apps/x-bot/requirements.txt
tweepy>=4.12.0
python-dotenv>=1.0.0
